Create README.md
The application populates all songs by fetching them from the SDcard. The system facilitates playing of songs, searching for songs alphabetically and also based on albums and artists. Songs can be added or deleted dynamically from playlists. The application can be run in the background and relevant notifications for the same have been provided.